==> notebooks/1.1_candidate_universe.ipynb <==
# Summary
This notebook builds the candidate universe of $p$-persistent arcs, for both single-scale and multiscale causal backbones. 

# Import libraries

import itertools
import jax
import jax.numpy as jnp
import matplotlib.pyplot as plt
import networkx as nx
import numpy as np
import pandas as pd
import sys

sys.path.insert(0, r"..\code")

from copy import deepcopy
from utils import save_obj,load_obj


data_dir="../data/"
ts_dir="../data/TimeSeriesAAL/" 
processed="../data/processed/"
diffreg="../data/processed/diff_regions/"

replace=False

# Names and index

#names
names = load_obj('region_names', processed)
idx_left = load_obj('index_left_regions',processed)
idx_right = load_obj('index_right_regions',processed)

names_left_df=names.iloc[idx_left].copy()
names_right_df=names.iloc[idx_right].copy()

names_left=names_left_df['Region']
names_right=names_right_df['Region']

names_reidx=names_left_df.copy()

names_reidx.reset_index(inplace=True, drop=True)
names_reidx.reset_index(inplace=True, drop=False)
names_reidx.set_index('Region', drop=True, inplace=True)

def retrieve_index(x, what):
    return names_reidx.loc[x[what]]['index']

# Persistent edges SS

if replace: writer = pd.ExcelWriter('1.1_single_scale_edges.xlsx')

threshold=80
alow, aup = (5, 95)
hemis=['right', 'left']
exams=[0, 1]
edges_df=pd.DataFrame()

for hemi,exam in itertools.product(hemis,exams):
    C_spk = load_obj('1.0_ss_{}_hemi_exam{}'.format(hemi,exam), processed).transpose((2,0,1))

    #occurrence
    C_occur=np.sign(np.abs(C_spk)).sum(axis=0)
    C_occur_df=pd.DataFrame(C_occur, index=names_right, columns=names_right)
    G=nx.from_pandas_adjacency(C_occur_df, create_using=nx.DiGraph)
    to_exclude=[(a,b) for a, b, attrs in G.edges(data=True) if attrs["weight"]<=threshold]
    [G.remove_edge(a,b) for a,b in to_exclude]
    G.remove_nodes_from(list(nx.isolates(G)))

    #percentiles
    C_percentile_alow=np.percentile(C_spk, alow, axis=0)
    C_percentile_aup=np.percentile(C_spk, aup, axis=0)

    C_percentile_alow_df=pd.DataFrame(C_percentile_alow, index=names_right, columns=names_right)
    C_percentile_aup_df=pd.DataFrame(C_percentile_aup, index=names_right, columns=names_right)

    [nx.set_edge_attributes(G, {(a,b):{'{}%'.format(alow):C_percentile_alow_df.loc[a,b], '{}%'.format(aup):C_percentile_aup_df.loc[a,b]}}) for a,b in G.edges] 

    if edges_df.empty:
        edges_df=nx.to_pandas_edgelist(G)
        edges_df['hemi']=hemi
        edges_df['day']=exam
    else:
        item=nx.to_pandas_edgelist(G)
        item['hemi']=hemi
        item['day']=exam
        edges_df=pd.concat((edges_df, item))

edges_df.rename(columns={'weight':'occurrence (out of 100)'}, inplace=True)

edges_df['count']=1.
edges_df_count = (edges_df[['source', 'target', 'count']].groupby(['source', 'target']).sum()).copy()
edges_df_grouped = (edges_df.groupby(['source', 'target', 'occurrence (out of 100)', '5%', '95%', 'hemi', 'day']).sum()).copy()
display(edges_df_count)
display(edges_df_grouped)

edges_df['source_idx']=edges_df.apply(lambda x: retrieve_index(x, 'source'), axis=1)
edges_df['target_idx']=edges_df.apply(lambda x: retrieve_index(x, 'target'), axis=1)

if replace:
    edges_df.to_excel(writer, sheet_name='Raw', index=True)
    edges_df_count.to_excel(writer, sheet_name='Count', index=True)
    edges_df_grouped.to_excel(writer, sheet_name='Grouped', index=True)

    writer.save()
    writer.close()

# Persistent edges MS

if replace: writer = pd.ExcelWriter('1.1_multi_scale_edges.xlsx')

J=5
N=len(names_right)
threshold=80
alow, aup = (5, 95)
hemis=['right', 'left']
exams=[0, 1]

for scale in range(J):
    edges_df=pd.DataFrame()
    for hemi,exam in itertools.product(hemis,exams):
        C_spk = load_obj('1.0_ms_{}_hemi_exam{}'.format(hemi,exam), processed).transpose((2,0,1))[:,scale*N:(scale+1)*N,scale*N:(scale+1)*N]

        #occurrence
        C_occur=np.sign(np.abs(C_spk)).sum(axis=0)
        C_occur_df=pd.DataFrame(C_occur, index=names_right, columns=names_right)
        G=nx.from_pandas_adjacency(C_occur_df, create_using=nx.DiGraph)
        to_exclude=[(a,b) for a, b, attrs in G.edges(data=True) if attrs["weight"]<=threshold]
        [G.remove_edge(a,b) for a,b in to_exclude]
        G.remove_nodes_from(list(nx.isolates(G)))

        #percentiles
        C_percentile_alow=np.percentile(C_spk, alow, axis=0)
        C_percentile_aup=np.percentile(C_spk, aup, axis=0)

        C_percentile_alow_df=pd.DataFrame(C_percentile_alow, index=names_right, columns=names_right)
        C_percentile_aup_df=pd.DataFrame(C_percentile_aup, index=names_right, columns=names_right)

        [nx.set_edge_attributes(G, {(a,b):{'{}%'.format(alow):C_percentile_alow_df.loc[a,b], '{}%'.format(aup):C_percentile_aup_df.loc[a,b]}}) for a,b in G.edges] 


        if edges_df.empty:
            edges_df=nx.to_pandas_edgelist(G)
            edges_df['hemi']=hemi
            edges_df['day']=exam
            edges_df['scale']=J-scale
        else:
            item=nx.to_pandas_edgelist(G)
            item['hemi']=hemi
            item['day']=exam
            item['scale']=J-scale
            edges_df=pd.concat((edges_df, item))

    edges_df.rename(columns={'weight':'occurrence (out of 100)'}, inplace=True)

    edges_df['count']=1.
    edges_df_count = (edges_df[['source', 'target', 'count']].groupby(['source', 'target']).sum()).copy()
    edges_df_grouped = (edges_df.groupby(['source', 'target', 'occurrence (out of 100)', '5%', '95%', 'hemi', 'day', 'scale']).sum()).copy()
    display(edges_df_count)
    display(edges_df_grouped)

    if replace:
        edges_df['source_idx']=edges_df.apply(lambda x: retrieve_index(x, 'source'), axis=1)
        edges_df['target_idx']=edges_df.apply(lambda x: retrieve_index(x, 'target'), axis=1)
        edges_df.to_excel(writer, sheet_name='Raw_scale{}'.format(J-scale), index=True)
        edges_df_count.to_excel(writer, sheet_name='Count_scale{}'.format(J-scale), index=True)
        edges_df_grouped.to_excel(writer, sheet_name='Grouped_scale{}'.format(J-scale), index=True)

if replace: 
    writer.save()
    writer.close()
